update - bug

+++ ViewModels/ChargerInfoViewModel.cs
@@ -11,10 +11,10 @@ namespace backend.ViewModels.Power
         public string charger_id { get; set; }
         public string chargergun_id { get; set; }
         public int charge_time { get; set; }
-        public int charge_current { get; set; }
-        public int charge_voltage { get; set; }
-        public int charge_kw { get; set; }
-        public int current_kw { get; set; }
+        public float charge_current { get; set; }
+        public float charge_voltage { get; set; }
+        public float charge_kw { get; set; }
+        public float current_kw { get; set; }
         public string soc { get; set; }
         public int trans_no { get; set; }
         public DateTime time { get; set; }
